Find the image digest with an unambiguous command
It seems that the build system docker retains the old tags and
RepoDigests for unchanged images: So inspecting an image tagged with
one repo will emit digests tagged for another repo as well.

Here, we work around this using the `docker images --format` command,
which emits only exactly the digest for the image, and construct the
correct fully-qualified digest name from the "name" input using it.

Also, emit a "digest" output now, which allows users to use the bare
digest (whatever they want to do with it).