Search repo digests for correct repo
If the same image is pushed to multiple repos, each repo may compute a
different digest (seriously). So, search through the list and find the
one that matches, and use *that* as the repo digest. This typically
comes up when creating a repo (e.g. an ECR repo) in the same Pulumi
program as pushing to it, and then `pulumi destroy && pulumi up` without
changing the Docker image will result in a local Docker image that has
repo digest for both the old deleted repo and the new repo.