Use the go SDK to perform docker actions instead of the CLI
This should eliminate all dependencies we have on the docker CLI,
replacing them with invocations of the docker API. Users only need to
have DOCKER_HOST set (or have docker available at the default docker
socket addresses).

It also makes all interactions we have with the docker API typed,
which is a nice win too.